Fix Nexus issue by StanlyHalo
For some odd reason, the _baks dictionary didn't contain a file which needed
deletion when stablizing the game. Added a check for key before attempting delete

+++ Api/Backups.cs
@@ -417,6 +417,9 @@ namespace MCC_Mod_Manager.Api {
             return !err;
         }
         public static bool DeleteBak(string path) {
+            if (!_baks.ContainsKey(path)) {
+                return false;
+            }
             if (Utility.DeleteFile(Config.Backup_dir + @"\" + _baks[path])) {
                 _baks.Remove(path);
                 return true;
